Handle non-git repos in Licensee::Project
If a project is detected to not be a git repo, simply scan the files in
that directory manually using FilesystemRepository substituting
Rugged::Repository.
