+++ lib/licensee.rb
@@ -8,6 +8,7 @@ require_relative "licensee/license"
 require_relative "licensee/licenses"
 require_relative "licensee/license_file"
 require_relative "licensee/project"
+require_relative "licensee/filesystem_repository"
 require_relative "licensee/matcher"
 require_relative "licensee/matchers/exact_matcher"
 require_relative "licensee/matchers/copyright_matcher"
